* Some internal changes/simplifications to the build system. * Fixed VS2008 x64 compilation.
git-svn-id: http://www.bci2000.org/svn/trunk@4509 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/SignalSource/Neuroscan/NeuroscanProtocol.h
@@ -155,11 +155,11 @@ class NscPacketHeader
       mDataSize( 0 )
     {
     };
-    NscPacketHeader( int id, int code, int value, int dataSize = 0 )
+    NscPacketHeader( int id, int code, int value, size_t dataSize = 0 )
     : mId( id ),
       mCode( code ),
       mValue( value ),
-      mDataSize( dataSize )
+      mDataSize( static_cast<int32_t>( dataSize ) )
     {
     };
     bool operator<( const NscPacketHeader& n ) const
